requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiofiles=22.1.0=pypi_0
aiosqlite=0.18.0=pypi_0
altair=4.2.2=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
attrs=22.2.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bleach=6.0.0=pypi_0
blinker=1.6.1=pypi_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=pypi_0
comm=0.1.3=pypi_0
debugpy=1.6.7=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.3=pypi_0
fqdn=1.5.1=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.2.1=pypi_0
ipykernel=6.22.0=pypi_0
ipython=8.12.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.6=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
json5=0.9.11=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.1.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=1.23.6=pypi_0
jupyter-server-fileid=0.9.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyter-server-ydoc=0.8.0=pypi_0
jupyter-ydoc=0.2.3=pypi_0
jupyterlab=3.6.3=pypi_0
jupyterlab-miami-nights=0.3.2=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.22.0=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc=7.2.0=h69d50b8_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
nbclassic=0.5.5=pypi_0
nbclient=0.7.3=pypi_0
nbconvert=7.3.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
nodejs=6.13.1=0
notebook=6.5.4=pypi_0
notebook-shim=0.2.2=pypi_0
openssl=1.1.1t=h7f8727e_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=23.0.1=py39h06a4308_0
platformdirs=3.2.0=pypi_0
plotly=5.14.1=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=11.0.0=pypi_0
pycparser=2.21=pypi_0
pydeck=0.8.0=pypi_0
pympler=1.0.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.9.16=h7a1cb2a_2
python-json-logger=2.0.7=pypi_0
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
qtconsole=5.4.2=pypi_0
qtpy=2.3.1=pypi_0
readline=8.2=h5eee18b_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4=pypi_0
sqlite=3.41.1=h5eee18b_0
stack-data=0.6.2=pypi_0
streamlit=1.21.0=pypi_0
tenacity=8.2.2=pypi_0
terminado=0.17.1=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=4.3=pypi_0
uri-template=1.2.0=pypi_0
urllib3=1.26.15=pypi_0
validators=0.20.0=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
wheel=0.38.4=py39h06a4308_0
widgetsnbextension=4.0.7=pypi_0
xz=5.2.10=h5eee18b_1
y-py=0.5.9=pypi_0
ypy-websocket=0.8.2=pypi_0
zlib=1.2.13=h5eee18b_0
